fix potential overflow in nodetool cfstats patch by Ryan King; reviewed by jbellis for CASSANDRA-2057
git-svn-id: https://svn.apache.org/repos/asf/cassandra/branches/cassandra-0.7@1063562 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/tools/NodeCmd.java
@@ -332,8 +332,8 @@ public class NodeCmd {
         {
             String tableName = entry.getKey();
             List<ColumnFamilyStoreMBean> columnFamilies = entry.getValue();
-            int tableReadCount = 0;
-            int tableWriteCount = 0;
+            long tableReadCount = 0;
+            long tableWriteCount = 0;
             int tablePendingTasks = 0;
             double tableTotalReadTime = 0.0f;
             double tableTotalWriteTime = 0.0f;
